Move boundary-violating files to their correct library tiers
Phase 2 boundary fixes: physically move files that had Core→IO or
Core→Algo back-dependencies to their proper library tiers.

Moves:
- FASTAContainer: DATASTRUCTURES → FORMAT (depends on FASTAFile)
- QTCluster: DATASTRUCTURES → ANALYSIS/MAPMATCHING (used only by QTClusterFinder)
- PosteriorErrorProbabilityModel: MATH/STATISTICS → ANALYSIS/ID (uses TextFile)
- OnDiscMSExperiment: KERNEL → FORMAT (uses FileHandler, IndexedMzMLHandler)

Compatibility redirect headers left at old locations so existing
#include directives and pyOpenMS .pxd files continue to work.

All 2455 tests pass (9 pre-existing external tool failures unchanged).
