+++ src/openms/algo_includes.cmake
@@ -52,9 +52,6 @@ if(NOT DISABLE_OPENSWATH)
 endif()
 include(${CMAKE_CURRENT_LIST_DIR}/source/APPLICATIONS/sources.cmake)
 
-# Add Core source files that were moved here due to Algo-level dependencies
-list(APPEND OpenMS_sources ${OpenMS_Core_moved_to_algo})
-
 set(OpenMS_Algo_sources ${OpenMS_sources} CACHE INTERNAL "OpenMS Algo source files")
 set(OpenMS_sources CACHE INTERNAL "")
 
@@ -101,9 +98,6 @@ if(NOT DISABLE_OPENSWATH)
   include(${CMAKE_CURRENT_LIST_DIR}/include/OpenMS/ANALYSIS/OPENSWATH/DATAACCESS/sources.cmake)
 endif()
 include(${CMAKE_CURRENT_LIST_DIR}/include/OpenMS/APPLICATIONS/sources.cmake)
-
-# Add Core headers that were moved here (parallel to source moves)
-list(APPEND OpenMS_sources_h ${OpenMS_Core_moved_to_algo_h})
 
 set(OpenMS_Algo_sources_h ${OpenMS_sources_h} CACHE INTERNAL "OpenMS Algo header files")
 set(OpenMS_sources_h CACHE INTERNAL "")
