+++ src/openms/core_includes.cmake
@@ -26,25 +26,22 @@ include(${CMAKE_CURRENT_LIST_DIR}/source/CHEMISTRY/MASSDECOMPOSITION/IMS/sources
 include(${CMAKE_CURRENT_LIST_DIR}/source/CHEMISTRY/MASSDECOMPOSITION/sources.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/source/IONMOBILITY/sources.cmake)
 
-# Remove source files that have IO/Algo back-dependencies and belong in those layers.
+# Remove source files that have IO back-dependencies and belong in the IO layer.
 # These files are leaf nodes in Core's dependency graph (no other Core .cpp depends on them)
-# but reference FORMAT types. They are compiled into OpenMS_IO or OpenMS_Algo instead.
+# but reference FORMAT types. They are compiled into OpenMS_IO instead.
+# Note: FASTAContainer, QTCluster, OnDiscMSExperiment, PosteriorErrorProbabilityModel
+# have been physically moved to their new directories in Phase 2.
 set(_core_to_io_sources
   source/CONCEPT/ClassTest.cpp        # test framework validation uses FORMAT file classes
   source/CONCEPT/FuzzyStringComparator.cpp  # uses TextFile
-  source/KERNEL/OnDiscMSExperiment.cpp      # uses FileHandler, PeakFileOptions, IndexedMzMLHandler
   source/METADATA/ID/IdentificationDataConverter.cpp  # uses MzTab types
   source/METADATA/ExperimentalDesign.cpp    # uses FileHandler, TextFile
 )
-set(_core_to_algo_sources
-  source/MATH/STATISTICS/PosteriorErrorProbabilityModel.cpp  # uses TextFile for gnuplot output
-)
-list(REMOVE_ITEM OpenMS_sources ${_core_to_io_sources} ${_core_to_algo_sources})
+list(REMOVE_ITEM OpenMS_sources ${_core_to_io_sources})
 
 # Capture accumulated sources into Core variable and reset accumulator
 set(OpenMS_Core_sources ${OpenMS_sources} CACHE INTERNAL "OpenMS Core source files")
 set(OpenMS_Core_moved_to_io ${_core_to_io_sources} CACHE INTERNAL "Core sources moved to IO")
-set(OpenMS_Core_moved_to_algo ${_core_to_algo_sources} CACHE INTERNAL "Core sources moved to Algo")
 set(OpenMS_sources CACHE INTERNAL "")
 
 ## Header files
@@ -67,20 +64,15 @@ include(${CMAKE_CURRENT_LIST_DIR}/include/OpenMS/CHEMISTRY/MASSDECOMPOSITION/IMS
 include(${CMAKE_CURRENT_LIST_DIR}/include/OpenMS/CHEMISTRY/MASSDECOMPOSITION/sources.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/include/OpenMS/IONMOBILITY/sources.cmake)
 
-# Remove headers for files moved to IO/Algo (parallel to source file moves above)
+# Remove headers for files compiled in IO (parallel to source file moves above)
 set(_core_to_io_headers
   include/OpenMS/CONCEPT/ClassTest.h
   include/OpenMS/CONCEPT/FuzzyStringComparator.h
-  include/OpenMS/KERNEL/OnDiscMSExperiment.h
   include/OpenMS/METADATA/ID/IdentificationDataConverter.h
   include/OpenMS/METADATA/ExperimentalDesign.h
 )
-set(_core_to_algo_headers
-  include/OpenMS/MATH/STATISTICS/PosteriorErrorProbabilityModel.h
-)
-list(REMOVE_ITEM OpenMS_sources_h ${_core_to_io_headers} ${_core_to_algo_headers})
+list(REMOVE_ITEM OpenMS_sources_h ${_core_to_io_headers})
 
 set(OpenMS_Core_sources_h ${OpenMS_sources_h} CACHE INTERNAL "OpenMS Core header files")
 set(OpenMS_Core_moved_to_io_h ${_core_to_io_headers} CACHE INTERNAL "Core headers moved to IO")
-set(OpenMS_Core_moved_to_algo_h ${_core_to_algo_headers} CACHE INTERNAL "Core headers moved to Algo")
 set(OpenMS_sources_h CACHE INTERNAL "")
