+++ src/openms/include/OpenMS/KERNEL/OnDiscMSExperiment.h
@@ -1,274 +1,3 @@
-// Copyright (c) 2002-present, OpenMS Inc. -- EKU Tuebingen, ETH Zurich, and FU Berlin
-// SPDX-License-Identifier: BSD-3-Clause
-//
-// --------------------------------------------------------------------------
-// $Maintainer: Hannes Roest $
-// $Authors: Hannes Roest $
-// --------------------------------------------------------------------------
-
+// Compatibility redirect — OnDiscMSExperiment moved to FORMAT module (library decomposition)
 #pragma once
-
-#include <OpenMS/INTERFACES/DataStructures.h>
-
-#include <OpenMS/KERNEL/MSExperiment.h>
-#include <OpenMS/KERNEL/MSSpectrum.h>
-#include <OpenMS/KERNEL/MSChromatogram.h>
-#include <OpenMS/METADATA/ExperimentalSettings.h>
-#include <OpenMS/FORMAT/HANDLERS/IndexedMzMLHandler.h>
-#include <OpenMS/FORMAT/OPTIONS/PeakFileOptions.h>
-
-#include <vector>
-#include <limits>
-
-#include <memory>
-
-namespace OpenMS
-{
-  /**
-    @brief Representation of a mass spectrometry experiment on disk.
-
-    @ingroup Kernel
-
-    This class allows random access to spectra and chromatograms in indexed mzML
-    files without loading the entire file into memory.
-
-    @section OnDiscMSExperiment_filtering Filtering with PeakFileOptions
-
-    PeakFileOptions can be used to filter data when retrieving spectra/chromatograms:
-    - RT range, MS level, and precursor m/z range filters: Checked BEFORE loading peak data (skips I/O)
-    - m/z range and intensity filters: Applied AFTER loading peak data
-
-    @note Unlike in-memory loading (FileHandler), where filtered spectra are completely
-    removed from the container, OnDiscMSExperiment preserves all indices. When a spectrum
-    doesn't pass RT range or MS level filters, getSpectrum() returns the spectrum with
-    metadata but no peaks (I/O is skipped). This preserves the index mapping to the file.
-
-    Example:
-    @code
-    OnDiscMSExperiment exp;
-    exp.openFile("data.mzML");
-    exp.getOptions().setMSLevels({2});  // Only want MS2
-    exp.getOptions().setRTRange(DRange<1>(100, 200));
-
-    for (Size i = 0; i < exp.size(); ++i)
-    {
-      MSSpectrum s = exp.getSpectrum(i);
-      if (s.empty()) continue;  // Filtered out, no I/O was performed
-      // Process MS2 spectrum in RT range...
-    }
-    @endcode
-
-    @note This implementation is @a not thread-safe since it keeps internally a
-    single file access pointer which it moves when accessing a specific
-    data item. Please provide a separate copy to each thread, e.g.
-
-    @code
-    #pragma omp parallel for firstprivate(ondisc_map)
-    @endcode
-
-  */
-  class OPENMS_DLLAPI OnDiscMSExperiment
-  {
-
-  typedef ChromatogramPeak ChromatogramPeakT;
-  typedef Peak1D PeakT;
-
-public:
-
-    /**
-      @brief Constructor
-
-      This initializes the object, use openFile to open a file.
-    */
-    OnDiscMSExperiment() = default;
-
-    /**
-      @brief Open a specific file on disk.
-
-      This tries to read the indexed mzML by parsing the index and then reading
-      the meta information into memory.
-
-      @return Whether the parsing of the file was successful (if false, the
-      file most likely was not an indexed mzML file)
-    */
-    bool openFile(const String& filename, bool skipMetaData = false);
-
-    /// Copy constructor
-    OnDiscMSExperiment(const OnDiscMSExperiment& source) :
-      filename_(source.filename_),
-      indexed_mzml_file_(source.indexed_mzml_file_),
-      meta_ms_experiment_(source.meta_ms_experiment_),
-      options_(source.options_)
-    {
-    }
-
-    /**
-      @brief Equality operator
-
-      This only checks whether the underlying file is the same and the parsed
-      meta-information is the same. Note that the file reader (e.g. the
-      std::ifstream of the file) might be in a different state.
-    */
-    bool operator==(const OnDiscMSExperiment& rhs) const
-    {
-      if (meta_ms_experiment_ == nullptr || rhs.meta_ms_experiment_ == nullptr) 
-      {
-        return filename_ == rhs.filename_ &&
-               meta_ms_experiment_ == rhs.meta_ms_experiment_;
-      }
-
-      // check if file and meta information is the same
-      return filename_ == rhs.filename_ &&
-             (*meta_ms_experiment_) == (*rhs.meta_ms_experiment_);
-      // do not check if indexed_mzml_file_ is equal -> they have the same filename...
-    }
-
-    /// Inequality operator
-    bool operator!=(const OnDiscMSExperiment& rhs) const
-    {
-      return !(operator==(rhs));
-    }
-
-    /**
-      @brief Checks if all spectra are sorted with respect to ascending RT
-
-      Note that we cannot check whether all spectra are sorted (except if we
-      were to load them all and check).
-    */
-    bool isSortedByRT() const
-    {
-      if (!meta_ms_experiment_) return false;
-
-      return meta_ms_experiment_->isSorted(false);
-    }
-
-    /// alias for getNrSpectra
-    inline Size size() const
-    {
-      return getNrSpectra();
-    }
-
-    /// returns whether spectra are empty
-    inline bool empty() const
-    {
-      return getNrSpectra() == 0;
-    }
-
-    /// get the total number of spectra available
-    inline Size getNrSpectra() const
-    {
-      return indexed_mzml_file_.getNrSpectra();
-    }
-
-    /// get the total number of chromatograms available
-    inline Size getNrChromatograms() const
-    {
-      return indexed_mzml_file_.getNrChromatograms();
-    }
-
-    /// returns the meta information of this experiment (const access)
-    std::shared_ptr<const ExperimentalSettings> getExperimentalSettings() const
-    {
-      return std::static_pointer_cast<const ExperimentalSettings>(meta_ms_experiment_);
-    }
-
-    std::shared_ptr<PeakMap> getMetaData() const
-    {
-      return meta_ms_experiment_;
-    }
-
-    /// alias for getSpectrum
-    inline MSSpectrum operator[](Size n)
-    {
-      return getSpectrum(n);
-    }
-
-    /**
-      @brief returns a single spectrum
-
-      If PeakFileOptions has m/z or intensity range set, the spectrum will be filtered accordingly.
-
-      @param[in] id The index of the spectrum
-    */
-    MSSpectrum getSpectrum(Size id);
-
-    /**
-      @brief returns a single spectrum (without applying PeakFileOptions filters)
-    */
-    OpenMS::Interfaces::SpectrumPtr getSpectrumById(Size id)
-    {
-      return indexed_mzml_file_.getSpectrumById((int)id);
-    }
-
-    /**
-      @brief returns a single chromatogram
-
-      If PeakFileOptions has RT or intensity range set, the chromatogram will be filtered accordingly.
-
-      @param[in] id The index of the chromatogram
-    */
-    MSChromatogram getChromatogram(Size id);
-
-    /**
-      @brief returns a single chromatogram
-
-      @param[in] id The native identifier of the chromatogram
-    */
-    MSChromatogram getChromatogramByNativeId(const std::string& id);
-
-    /**
-      @brief returns a single spectrum
-
-      @param[in] id The native identifier of the spectrum
-    */
-    MSSpectrum getSpectrumByNativeId(const std::string& id);
-
-    /**
-      @brief returns a single chromatogram
-    */
-    OpenMS::Interfaces::ChromatogramPtr getChromatogramById(Size id);
-
-    /// sets whether to skip some XML checks and be fast instead
-    void setSkipXMLChecks(bool skip);
-
-    /// @brief Mutable access to the options for loading/storing
-    PeakFileOptions& getOptions();
-
-    /// @brief Non-mutable access to the options for loading/storing
-    const PeakFileOptions& getOptions() const;
-
-    /// @brief set options for loading/storing
-    void setOptions(const PeakFileOptions& options);
-
-private:
-
-    /// Private Assignment operator -> we cannot copy file streams in IndexedMzMLHandler
-    OnDiscMSExperiment& operator=(const OnDiscMSExperiment& /* source */);
-
-    void loadMetaData_(const String& filename);
-
-    MSChromatogram getMetaChromatogramById_(const std::string& id);
-
-    MSSpectrum getMetaSpectrumById_(const std::string& id);
-
-protected:
-
-    /// The filename of the underlying data file
-    String filename_;
-    /// The index of the underlying data file
-    Internal::IndexedMzMLHandler indexed_mzml_file_;
-    /// The meta-data
-    std::shared_ptr<PeakMap> meta_ms_experiment_;
-    /// Mapping of chromatogram native ids to offsets
-    std::unordered_map< std::string, Size > chromatograms_native_ids_;
-    /// Mapping of spectra native ids to offsets
-    std::unordered_map< std::string, Size > spectra_native_ids_;
-    /// Options for loading / storing
-    PeakFileOptions options_;
-  };
-
-typedef OpenMS::OnDiscMSExperiment OnDiscPeakMap;
-
-} // namespace OpenMS
-
-
+#include <OpenMS/FORMAT/OnDiscMSExperiment.h>
