+++ src/openms/include/OpenMS/KERNEL/sources.cmake
@@ -26,7 +26,6 @@ MRMTransitionGroup.h
 MSChromatogram.h
 MSExperiment.h
 MSSpectrum.h
-OnDiscMSExperiment.h
 Peak1D.h
 Peak2D.h
 PeakIndex.h
