+++ src/openms/include/OpenMS/MATH/STATISTICS/sources.cmake
@@ -11,7 +11,6 @@ GumbelMaxLikelihoodFitter.h
 Histogram.h
 KernelDensityEstimation.h
 MultipleTesting.h
-PosteriorErrorProbabilityModel.h
 RankData.h
 )
